October 28th work note
As it stands, I was able to get the scene to reappear and fix the black screen issue. Changing the location of the loadGUI() method fixed it. However, neither the sample button or my ImageOnly button will show, and the quotes seem to get blocked as well as a sidenote. The console log in the Zen Garden tab says its an issue with "AdvancedDynamicTexture", which we pinpointed Wednesday, however, despite adding extra GUI libraries and altering the code in multiple ways, it cannot seem to find it, therefore reading it as "undefined". I've looked through forums with similar errors, and nothing consistent is there.  These are just notes for me to go back to during 10/29 morning meeting.